Fix `uinput-echo` example
In the setup of the device, the keys were initialized to

```
[Codes.KeyA .. Codes.KeyZ] ++ [Codes.Key1 .. Codes.Key0]
```

But that is incorrect: `KeyA` corresponds to value 30 and `KeyZ`
corresponds to value 44; this therefore excludes keys such as `KeyB`
(48) and others. This can be observed by running the example and typing
`abcde`; only `ad` will be echoed back.

Perhaps the better fix would be to modify the `Enum` instance, but that
might result in difficult to debug backwards incompatibility problems,
so this PR simply fixes the example.